cleanup and break things
