+++ app/items/[uuid]/page.tsx
@@ -1,9 +1,11 @@
 import React from "react";
 import { Metadata } from "next";
 import PageLayout from "../../src/components/pageLayout/pageLayout";
-import Item from "../../src/components/items/item";
-import { getItemData, stringToSlug } from "../../src/utils/utils";
+// import Item from "../../src/components/items/item";
+import { stringToSlug } from "../../src/utils/utils";
+import { getItemData } from "../../src/utils/api";
 import { ItemModel } from "../../src/models/item";
+import { ItemPage } from "../../src/components/pages/itemsPage/itemPage";
 
 type ItemProps = {
   params: {
@@ -14,6 +16,7 @@ type ItemProps = {
 
 const getItemModel = async (uuid: string) => {
   const data = await getItemData(uuid);
+  // console.log("data is: ", data)
   const item = new ItemModel(data);
   return item;
 };
@@ -28,23 +31,24 @@ export async function generateMetadata({
   };
 }
 
-export default async function ItemPage({ params }: ItemProps) {
-  const item = await getItemModel(params.uuid);
-  const pageName = `items|${stringToSlug(item.title)}`;
+export default async function Item({ params }: ItemProps) {
+  // const item = await getItemModel(params.uuid);
+  // const pageName = `items|${stringToSlug(item.title)}`;
   return (
-    <PageLayout
-      activePage="item"
-      breadcrumbs={[
-        { text: "Home", url: "/" },
-        { text: "All Items", url: "/items" },
-        {
-          text: `${item.title}`,
-          url: `/items/${params.uuid}`,
-        },
-      ]}
-      adobeAnalyticsPageName={pageName}
-    >
-      <Item item={item} />
-    </PageLayout>
+    // <PageLayout
+    //   activePage="item"
+    //   breadcrumbs={[
+    //     { text: "Home", url: "/" },
+    //     { text: "All Items", url: "/items" },
+    //     {
+    //       text: `${item.title}`,
+    //       url: `/items/${params.uuid}`,
+    //     },
+    //   ]}
+    //   adobeAnalyticsPageName={pageName}
+    // >
+    //   <Item item={item} />
+    // </PageLayout>
+    <ItemPage params={params} />
   );
 }
