+++ package.json
@@ -46,6 +46,7 @@
     "ts-node": "10.9.1",
     "typescript": "4.9.5",
     "ua-parser-js": "^1.0.38",
+    "uuid": "^10.0.0",
     "vite": "4.5.2",
     "winston": "3.9.0"
   },
